amp/madt: MADT Type 9 + 32-bit x2APIC IDs (S2, real-HW big-iron)
amp_collect_lapics() now parses BOTH MADT Type 0 (Local APIC, 8-bit id) and
Type 9 (Local x2APIC, 32-bit id), deduped. Logical processors with APIC id
>= 255 are listed only via Type 9 (ACPI 6.5 5.2.12.12), so the previous
Type-0-only collector silently dropped every such core on a >254-LP machine.

Widened the topology to full 32-bit ids end to end:
 - CoreDescriptor.lapic_id, AmpLayout.bsp_lapic_id, PerCoreData.lapic_id,
   amp_collect_lapics() ids and amp_init()'s working set -> uint32_t.
 - lapic_send_ipi() dest -> uint32_t (lapic_icr_write already delivers the
   full 32-bit destination on the x2APIC path; xAPIC uses dest[7:0]).
 - Replaced the 256-entry lapic_to_index[] reverse map (cannot index a 32-bit
   id) with amp_index_for_lapic(), a width-independent linear scan used only on
   slow paths (amp_init build + amp_get_core_index early-boot fallback).

Verified clean build; BIOS/UEFI x 1c/4c boot to shell, 4 cores detected, all
APs online, interactive shell OK. (The Type 9 path needs a >254-LP machine to
exercise and is validated by spec + review; QEMU enumerates Type 0.)

+++ src/kernel/arch/x86-64/amp/amp.h
@@ -7,7 +7,7 @@
 #define AP_TRAMPOLINE_PHYS  0x8000
 
 typedef struct {
-    uint8_t  lapic_id;
+    uint32_t lapic_id;        // xAPIC (8-bit) or x2APIC (32-bit) ID — full width
     uint8_t  core_index;
     bool     is_bsp;
     bool     is_kcore;
@@ -20,7 +20,7 @@ typedef struct {
     uint8_t  k_count;
     uint8_t  app_count;
     uint8_t  bsp_index;
-    uint8_t  bsp_lapic_id;
+    uint32_t bsp_lapic_id;    // full-width APIC ID of the BSP
     bool     multicore_active;
 } AmpLayout;
